requirements/gemma3-4b_requirements.txt
torchvision
jupyter
numpy==1.26.4
matplotlib
commitizen
accelerate
transformers
scipy